pyproject.toml
[build-system]
requires=["setuptools"]
build-backend="setuptools.build_meta"

[project]
name="corefacility"
version="0.1.1"
dependencies=[
    "psutil",
    "python-dotenv",
    "django",
    "django-configurations",
    "djangorestframework",
    "pillow",
    "parameterized",
    "pytz",
    "colorama",
    "urllib3",
    "bs4",
    "dateutils",
    "numpy",
    "matplotlib",
    "scipy",
    "pandas",
]

[project.scripts]
corefacility="ru.ihna.kozhukhov.corefacility.cli:main"
corefacility-daemon="ru.ihna.kozhukhov.corefacility.daemon:main"

[tools.setuptools]
include-package-data=true

[tools.setuptools.packages.find]
where=["src"]
